Functioning App 4 (new image)
The link I originally had for my new image was a google link, which didn't show an image at all in the app. I went back and found the image link from the original site and inserted that instead. I also adjusted the "show results" line of code and moved it to instead the "div" tags to see if that would keep the "Results" header in place and not make it disappear.