Functioning App 4 (new image)
The first picture that I adjusted now resizes on my phone, so I tried a new, more square picture for my second one. I also changed the "show results" code back to the way it was before because it did not work when I moved it in between the "div" tags.