changes to style formatting
I got rid of the code for the bar graph as it did not work and replaced it with the previous code I had from app 4 (that displayed the results in percentages). I also changed the colors from purple to a blue that is not as obnoxious and more appealing to the eyes.