Final App 7- comparing symptoms of covid-19, ,cold, flu, and allergies
"I went through and changed the results page back to its original form, because while it may be unappealing, at least it works and I can show something for my results. I also fixed some of the colors that were off because the questionnaire looked unprofessional before. I went ahead and added some more comments as well, especially about a potential different way to display the results."
I completed all of these things in my last commit^, but I went through the survey one more time and found a tiny mistake in the header of one of my questions, so I am committing these changes to fix that and officially complete this app.