Final App 7- comparing symptoms of covid-19, ,cold, flu, and allergies
I had originally taken out my failed attempt of displaying results as a bar graph in the canvas because it was a lot of code that made the app look cluttered and was overwhelming to scroll through. However, I just noticed that Dr. Riedel asked us to document it anyways to show our attempts, so I went back and added the code I had tried to tweak and function.